refactor dataset generation and include feedback in dataset episode construction
