+++ scripts/make_demo_with_feedback.py
@@ -1,3 +1,4 @@
+from src.dataset.custom_dataset import CustomDataset
 from src.utils.argparsing import get_args
 from src.utils.demos import (
     DEFAULT_EASY_ACTIONS,
@@ -8,13 +9,10 @@ from src.utils.demos import (
     DEFAULT_HARD_SEED,
     DemoVideo,
 )
-from generate_datasets import get_dataset
 
 if __name__ == "__main__":
     args = get_args()
-    args["load_dataset_if_exists"] = False
-    args["demo"] = "custom"
-    args["demo_episode"] = 0
+    args["demo"] = "from_default_easy"
     if args["demo"] == "from_default_hard":
         config = DEFAULT_HARD_CONFIG
         seed = DEFAULT_HARD_SEED
@@ -24,9 +22,9 @@ if __name__ == "__main__":
         seed = DEFAULT_EASY_SEED
         actions = DEFAULT_EASY_ACTIONS
     else:
-        dataset = get_dataset(args)
+        dataset = CustomDataset(args).get_dataset()
         config = dataset.environment_name
         seed = dataset.seed_used
         actions = dataset.episodes[args["demo_episode"]].actions
-    # demo = DemoVideo(config, seed, actions, args["demo"], args["output_dir"])
-    # demo.make_demo_video()
+    demo = DemoVideo(config, seed, actions, args["demo"], args["output_dir"])
+    demo.make_demo_video()
